raw conversion of Property specs
